add BM25

+++ index/src/main/java/it/unipi/dii/mircv/index/algorithms/Spimi.java
@@ -18,7 +18,7 @@ public class Spimi {
     private Logs log;
     private int indexCounter;
     private int documentCounter;
-    private int totDocLength;
+    private long totDocLength;
 
     public Spimi(String collection) {
         this.COLLECTION_PATH = collection;
